refactor: 💡 Removed robot specific constants
