+++ raptacon3200/constants/common.py
@@ -1,0 +1,3 @@
+"""
+Collection of constants that are used across all of the robots
+"""
