--- raptacon3200/constants/swerve.py
@@ -1,110 +1,0 @@
-"""
-Collection of numeric constants that define physical properties of arobot
-"""
-
-# Native imports
-import math
-
-#############################
-# ROBOT ###################
-#############################
-
-
-class RobotConstants:
-    massKG: float = 74.088
-    MOI: float = 6.883
-
-
-#############################
-# SENSORS ###################
-#############################
-
-
-class SparkMaxConstants:
-    faultRateMs: int = 50
-    motorPosRateMs: int = 20
-    appliedOutputRateMs: int = 10
-    motorTelmRateMs: int = 50
-    analogRateMs: int = 1833
-    altEncoderRateMs: int = 1050
-    dutyCycleEncRateMs: int = 2150
-    dutyCycleEncVelRateMs: int = 3150
-
-
-#############################
-# SWERVE ###################
-#############################
-
-
-class SwerveDriveConsts(RobotConstants):
-    moduleFrontLeftX: float = 0.31115
-    moduleFrontLeftY: float = 0.26035
-    moduleFrontRightX: float = 0.31115
-    moduleFrontRightY: float = -0.26035
-    moduleBackLeftX: float = -0.31115
-    moduleBackLeftY: float = 0.26035
-    moduleBackRightX: float = -0.31115
-    moduleBackRightY: float = -0.26035
-
-    invertGyro: bool = False
-    moduleFrontLeftInvertDrive: bool = True
-    moduleFrontRightInvertDrive: bool = True
-    moduleBackLeftInvertDrive: bool = True
-    moduleBackRightInvertDrive: bool = True
-
-    moduleFrontLeftInvertSteer: bool = True
-    moduleFrontRightInvertSteer: bool = True
-    moduleBackLeftInvertSteer: bool = True
-    moduleBackRightInvertSteer: bool = True
-
-    maxTranslationMPS: float = 4.14528
-    maxAngularDPS: float = math.degrees(maxTranslationMPS / math.hypot(moduleFrontLeftY, moduleFrontLeftX))
-
-
-class SwerveModuleMk4iConsts(SwerveDriveConsts):
-    """
-    https://github.com/SwerveDriveSpecialties/swerve-lib/blob/develop/src/main/java/com/swervedrivespecialties/swervelib/ctre/Falcon500DriveControllerFactoryBuilder.java
-    """
-    kNominalVoltage: float = 12.0
-    kDriveCurrentLimit: int = 40
-    kSteerCurrentLimit: int = 20
-    kRampRate: float = 0.25
-    kTicksPerRotation: int = 1
-    kCanStatusFrameHz: int = 10
-    quadratureMeasurementRateMs: int = 10
-    quadratureAverageDepth: int = 2
-    numDriveMotors: int = 1
-    motorType: str = "NEO" # should be an option in wpimath.system.plant.DCMotor
-
-
-class SwerveModuleMk4iL1Consts(SwerveModuleMk4iConsts):
-    """
-    https://docs.yagsl.com/configuring-yagsl/standard-conversion-factors
-    """
-    wheelDiameter: float = 0.10033
-    driveGearRatio: float = 8.14
-    steerGearRatio: float = 150 / 7
-
-    drivePositionConversionFactor : float = (math.pi * wheelDiameter) / (driveGearRatio * SwerveModuleMk4iConsts.kTicksPerRotation)
-    driveVelocityConversionFactor: float = drivePositionConversionFactor / 60.0
-    steerPositionConversionFactor: float = 360 / (steerGearRatio * SwerveModuleMk4iConsts.kTicksPerRotation)
-    steerVelocityConversionFactor: float = steerPositionConversionFactor / 60.0
-
-    moduleType: str = "Mk4i_L1"
-
-
-class SwerveModuleMk4iL2Consts(SwerveModuleMk4iConsts):
-    """
-    https://docs.yagsl.com/configuring-yagsl/standard-conversion-factors
-    """
-    wheelDiameter: float = 0.10033
-    wheelCOF: float = 1.0
-    driveGearRatio: float = 6.75
-    steerGearRatio: float = 150 / 7
-
-    drivePositionConversionFactor : float = (math.pi * wheelDiameter) / (driveGearRatio * SwerveModuleMk4iConsts.kTicksPerRotation)
-    driveVelocityConversionFactor: float = drivePositionConversionFactor / 60.0
-    steerPositionConversionFactor: float = 360 / (steerGearRatio * SwerveModuleMk4iConsts.kTicksPerRotation)
-    steerVelocityConversionFactor: float = steerPositionConversionFactor / 60.0
-
-    moduleType: str = "Mk4i_L2"
